requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
cycler==0.11.0
# Commented for Python 3.8:
#dataclasses>=0.8
decorator==4.4.2
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.3
google-auth==2.16.3
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
idna-ssl==1.1.0
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.3.1
lxml==4.9.2
Markdown==3.3.7
matplotlib==3.3.4
MedPy==0.4.0
multidict==5.2.0
natsort==8.2.0
networkx==2.5.1
nibabel==3.2.2
numpy==1.19.5
oauthlib==3.2.2
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pydicom==2.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2023.2
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
SimpleITK==2.1.1.2
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2020.9.3
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.10.2+cu113
torchmetrics==0.8.2
--extra-index-url https://download.pytorch.org/whl/cu113
torchvision==0.11.3+cu113
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.15
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
sqlalchemy==1.4.47
# Commented for Python 3.8:
# lzo
# => Install the lzo library and its development headers.
# python-lzo==1.14
mat73==0.60
